Move fault tolerant tests to separate module
Commit moves fault tolerant tests out of trino-hive, trino-iceberg,
trino-delta to new trino-faulttolerant-test module.
With fault tolerance tests in connector plugins those need to depend (in
test scope) on trino-exchange-filesystem. This proved to be a problem
and resulted in influencing compile classpath of plugins in a way that
connectors were failing at runtime (some needed transitive dependecies
were rescoped from compile to test).

See: https://github.com/trinodb/trino/issues/12674
